Adding results for Philippines data.
Adding results for Philippines data analyzed under the old msBayes
implementation. The data were analyzed with a uniform prior on
divergence time that has the same variance as the exponential prior used
in the analyses under the new implementation. A uniform prior on theta
was selected that has the same mean as the exponential used in the
analyses under the new implementation.