Making some edits to dpp ms in proofing stage.
Adding Table S1 to supporting information that give an example of the
divergence model sample space when Y=4 taxa.

Also, making some edits to the method section where the sample space of
ordered/unordered divergence models is introduced. Making the language
in this section more exact and introducing my loose use of "ordered" and
"unordered" subsequent to this section.