Updating plots for empirical Philippine results.
Modifying `empirical-analyses/bin/plot-results.py` to plot dpp-msbayes
vs msbayes psi results with the priors included. Also committing updated
PDFs of other plots produced by this script, but the code to produce
these plots was not changed (slight differences in fonts due to
different versions of PyMsBayes and/or matplotlib).